# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Inventory API",
        "description": "API documentation for managing a book inventory",
        "version": "1.0.0"
    },
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Great Gatsby"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "F. Scott Fitzgerald"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-3-16-148410-0"
                                    },
                                    "published_year": {
                                        "type": "integer",
                                        "example": 1925
                                    },
                                    "pages": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "English"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    }
                }
            }
        },
        "/books/create": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Show form to create a new book",
                "operationId": "95002cbbe04c3d3f83a7a3ccd7d1dcce",
                "responses": {
                    "200": {
                        "description": "Form for creating a new book"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Display a specific book",
                "operationId": "e94109dcb4f4d40b58a210b43d94016d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update an existing book",
                "operationId": "7430c1078c3cb5da552b2fd50409f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Great Gatsby"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "F. Scott Fitzgerald"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-3-16-148410-0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a specific book",
                "operationId": "41c44a46960ad1a921ba5bfbf16ea7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    }
                }
            }
        },
        "/books/{id}/edit": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Show form to edit an existing book",
                "operationId": "5dd8e95ce9f2e28c659b8e917b8b42d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form for editing a book"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "required": [
                    "title",
                    "author",
                    "isbn"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "author": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-3-16-148410-0"
                    },
                    "publisher": {
                        "type": "string",
                        "example": "Charles Scribner's Sons"
                    },
                    "published_year": {
                        "type": "integer",
                        "example": 1925
                    },
                    "pages": {
                        "type": "integer",
                        "example": 180
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T14:48:00.000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T14:48:00.000Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}